Add GLIF_CONFIRMATIONS env var
Primarily for testing, to allow reducing the number of confirmations
from 5 to 1, so tests can complete more quickly.

A more comprehensive approach might be to accept the number of
confirmations as a parameters, so the that the cli could have
a command flag to set it. For now, just using an environment
variable was the lowest impact way to enable testing.